various warning fixes for unsigned, signed conversions and similar things
git-svn-id: https://scm.projects.hlrs.de/svn/ls1/autopas/trunk@5978 a63bd714-7e14-4b5e-94e7-302c8c8ff188

+++ src/cells/FullParticleCell.h
@@ -22,7 +22,7 @@ public:
 	void addParticle(Particle& m) override {
 		_mols.push_back(m);
 	}
-	int numParticles() const override {return _mols.size();}
+	unsigned long numParticles() const override {return _mols.size();}
 	bool isNotEmpty() const override { return numParticles() > 0 ; }
 	std::vector<Particle> _mols;
 };
