+++ src/cells/ParticleCell.h
@@ -16,7 +16,7 @@ public:
 	virtual ~ParticleCell() = default;
 	virtual void addParticle(Particle& p) = 0;
 	virtual void moleculesAt(int i, Particle *& rmm_or_not_pointer) = 0; // TODO: consider making return type Particle*
-	virtual int numParticles() const = 0;
+	virtual unsigned long numParticles() const = 0;
 	virtual bool isNotEmpty() const = 0;
 };
